+++ src/components/common/index.ts
@@ -1,5 +1,5 @@
 export { default as Card } from './Card';
 export { default as CategorySearch } from './CategorySearch';
 export { default as Selectbox } from './Selectbox';
-export { default as SearchForm } from './searchForm/SearchForm';
+export { default as SearchOverview } from './searchOverview/SearchOverview';
 export { default as SearchInput } from './SearchInput';
